integrating GEMFile with tikaextraction
